Laba11/requirements.txt
Django==4.2.0
python-telegram-bot==21.4
python-dotenv==1.0.1
